🛠️ [refactor] Modify demo storage in the apps directory to packages
