+++ apps/admin/src/pages/demo/page/analysis/index.vue
@@ -1,0 +1,6 @@
+<template>
+  <Analysis />
+</template>
+<script setup>
+import { Analysis } from '@vben/demo'
+</script>
