+++ packages/demo/index.ts
@@ -16,3 +16,11 @@ export { default as Time } from './src/Time/index.vue'
 export { default as BasicScroll } from './src/Scroll/basic.vue'
 export { default as ActionScroll } from './src/Scroll/action.vue'
 export { default as VirtualScroll } from './src/Scroll/virtual.vue'
+
+export { default as Analysis } from './src/Analysis/index.vue'
+export { default as Workbench } from './src/Workbench/index.vue'
+export { default as Account } from './src/Account/index.vue'
+export { default as Department } from './src/Department/index.vue'
+export { default as Menu } from './src/Menu/index.vue'
+export { default as Password } from './src/Password/index.vue'
+export { default as Role } from './src/Role/index.vue'
