+++ packages/demo/package.json
@@ -15,12 +15,14 @@
   "dependencies": {
     "@amap/amap-jsapi-loader": "^1.0.1",
     "@tinymce/tinymce-vue": "^5.1.1",
+    "@vben/components": "workspace:*",
     "@vben/hooks": "workspace:*",
     "@vben/request": "workspace:*",
     "@vben/utils": "workspace:*",
-    "@vben/components": "workspace:*",
     "@vben/vbencomponents": "workspace:*",
-    "vue": "^3.3.6"
+    "echarts": "^5.4.3",
+    "vue": "^3.3.6",
+    "vue-echarts": "^6.6.1"
   },
   "devDependencies": {
     "@amap/amap-jsapi-types": "^0.0.13",
