+++ packages/demo/src/Workbench/index.vue
@@ -374,4 +374,3 @@ const memberOption = ref({
   visibility: visible;
 }
 </style>
-zu
